+++ userservice/src/main/java/com/micro/userservice/repo/CartItemRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.micro.userservice.entity.CartItem;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem, Long> { }
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> { }
